% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tda.R
\name{diagram}
\alias{diagram}
\title{Compute persistence diagram of data.}
\usage{
diagram(input, type, dim_max = 3, threshold = Inf, modulus = 2,
  do_cocycles = 0)
}
\arguments{
\item{input}{Either a distance matrix or a point cloud matrix.}

\item{type}{'point-cloud' or 'distance-matrix'}

\item{threshold}{Maximum distance to generate edge in simplicial complex.}

\item{modulus}{Homology coefficients.}

\item{do_cocycles}{Set to 1 to generate cocycles, 0 otherwise.}
}
\value{
The persistence data, including persistence pairs.
}
\description{
Compute persistence diagram of data.
}
