% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tda.R
\name{landscape}
\alias{landscape}
\title{Compute landscape of a persistence diagram.}
\usage{
landscape(PersistenceDiagram, degree = NULL, exact = FALSE, dx = 0.1,
  min_x = 0, max_x = 10, threshold = -1)
}
\arguments{
\item{PersistenceDiagram}{A pd object from the diagram function, or a list of pairs.}

\item{degree}{If input is pd object then degree specifies which degree to select from.}

\item{exact}{Set to TRUE for exact computation, FALSE for discrete.}

\item{dx}{Domain grid diameter for discrete PL.}

\item{max_x}{Domain threshold for discrete PL.}

\item{threshold}{Threshold used to compute PD (could be inf).}
}
\value{
A PersistenceLandscape object.
}
\description{
Compute landscape of a persistence diagram.
}
