% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tda.R
\name{PLplot}
\alias{PLplot}
\title{Plot a persistenceLandscape object.}
\usage{
PLplot(PersistenceLandscape, infinity_sub = -1)
}
\arguments{
\item{PersistenceLandscape}{A PL object.}

\item{infinity_sub}{When using an exact representation, a value of Inf can appear in the PL. infinity_sub replaces
this with a different value in the plot.}
}
\description{
Plot a persistenceLandscape object.
}
